pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kalshi-trading"
version = "0.1.0"
description = "ESPN-informed Kalshi event contract trading system"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Your Name", email = "you@example.com" }
]
dependencies = [
    "httpx>=0.27.0",           # Async HTTP client for API calls
    "pydantic>=2.0",           # Data validation and settings
    "pydantic-settings>=2.0",  # Environment-based config
    "cryptography>=42.0",      # RSA signing for Kalshi auth
    "pyyaml>=6.0",             # Strategy config files
    "structlog>=24.0",         # Structured logging
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
    "respx>=0.21",             # Mock httpx requests
    "ruff>=0.5",               # Linting and formatting
    "mypy>=1.10",              # Type checking
]

[project.scripts]
kalshi-trading = "kalshi_trading.cli:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "integration: marks tests requiring live API access",
    "slow: marks tests as slow",
]

[tool.coverage.run]
source = ["src/kalshi_trading"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
